Move Encode types to exported module
