+++ objc2-foundation/examples/custom_class.rs
@@ -2,10 +2,10 @@ use std::ptr::NonNull;
 use std::sync::Once;
 
 use objc2::declare::ClassDecl;
+use objc2::encode::{Encoding, RefEncode};
 use objc2::rc::{Id, Owned};
 use objc2::runtime::{Class, Object, Sel};
-use objc2::{msg_send, sel};
-use objc2::{Encoding, Message, RefEncode};
+use objc2::{msg_send, sel, Message};
 use objc2_foundation::{INSObject, NSObject};
 
 /// In the future this should be an `extern type`, if that gets stabilized,
