+++ objc2/src/message/apple/mod.rs
@@ -1,4 +1,5 @@
-use super::{conditional_try, Encode, MessageArguments, MessageError};
+use super::{conditional_try, MessageArguments, MessageError};
+use crate::encode::Encode;
 use crate::ffi;
 use crate::runtime::{Class, Imp, Object, Sel};
